requirements.txt
flask>=3.0.0
groq>=1.0.0
python-dotenv>=1.0.0
requests>=2.30.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
joblib>=1.3.0
xgboost>=2.0.0
scikit-learn>=1.3.0
